Move WikiTransformation to new module Wikitext.
