+++ web/php/Modules/Account/PM/PMPreviewModule.php
@@ -7,7 +7,7 @@ namespace Wikidot\Modules\Account\PM;
 
 use Wikidot\DB\PrivateMessage;
 use Wikidot\Utils\AccountBaseModule;
-use Wikidot\Utils\WikiTransformation;
+use Wikidot\Wikitext\WikiTransformation;
 
 class PMPreviewModule extends AccountBaseModule
 {
