+++ web/php/Modules/Forum/FrontForumModule.php
@@ -15,7 +15,7 @@ use Ozone\Framework\SmartyModule;
 use Wikidot\Utils\ProcessException;
 use Wikidot\Utils\WDRenderUtils;
 use Wikidot\Utils\WDStringUtils;
-use Wikidot\Utils\WikiTransformation;
+use Wikidot\Wikitext\WikiTransformation;
 
 class FrontForumModule extends SmartyModule
 {
