+++ web/php/Modules/Forum/Sub/ForumPostRevisionModule.php
@@ -7,7 +7,7 @@ use Wikidot\DB\ForumPostRevisionPeer;
 
 use Ozone\Framework\SmartyModule;
 use Wikidot\Utils\ProcessException;
-use Wikidot\Utils\WikiTransformation;
+use Wikidot\Wikitext\WikiTransformation;
 
 class ForumPostRevisionModule extends SmartyModule
 {
