+++ web/php/Modules/History/PageVersionModule.php
@@ -6,7 +6,7 @@ use Wikidot\DB\PageRevisionPeer;
 
 use Ozone\Framework\SmartyModule;
 use Wikidot\Utils\ProcessException;
-use Wikidot\Utils\WikiTransformation;
+use Wikidot\Wikitext\WikiTransformation;
 
 class PageVersionModule extends SmartyModule
 {
